properly interaction between the scenes
