--- Assets/Scenes/SampleSceneSettings.lighting.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ae9ac2215aebd0b4f879a69a304f4b06
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 4890085278179872738
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
