--- Assets/Scripts/CreditsThread.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 95fe608da9b66c24a979615bc8057a68
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
