+++ Assets/Scripts/SessionThread.cs
@@ -3,22 +3,16 @@ using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UI;
 using UnityEngine.XR.ARFoundation;
+using UnityEngine.SceneManagement;
 
 public class SessionThread : MonoBehaviour
 {
-    [SerializeField]
-    private GameObject Interface_Panel;
-
-    [SerializeField]
-    public Button btn_Back;
-
     private void Awake()
     {
-        btn_Back.onClick.AddListener(Back);
     }
-
-    private void Back()
+    
+    public void BackToMenu()
     {
-        Interface_Panel.gameObject.SetActive(true);
+        SceneManager.LoadScene(0);
     }
 }
