+++ ProjectSettings/EditorBuildSettings.asset
@@ -6,8 +6,11 @@ EditorBuildSettings:
   serializedVersion: 2
   m_Scenes:
   - enabled: 1
-    path: Assets/Scenes/UI Test.unity
-    guid: b6b7b99b0e5560e4f981baedc3bca052
+    path: Assets/Scenes/UI.unity
+    guid: 4c805d205f52f7942b8a3fb5919cfc42
+  - enabled: 1
+    path: Assets/Scenes/Tracking.unity
+    guid: 0cb39c708883ed94c8ea213016f278ba
   m_configObjects:
     UnityEditor.XR.ARCore.ARCoreSettings: {fileID: 11400000, guid: cb5b441f55ce4fe44871af9ef178bc3b, type: 2}
     com.unity.xr.management.loader_settings: {fileID: 11400000, guid: 24379f5cfc162124cbdf0ddb17657bba, type: 2}
