+++ ProjectSettings/TagManager.asset
@@ -8,10 +8,10 @@ TagManager:
   - Default
   - TransparentFX
   - Ignore Raycast
-  - UI_Credits
+  - 
   - Water
   - UI
-  - Session
+  - 
   - 
   - 
   -
